# BACKEND/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "cuidarteplus API (español)",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:4444" }],
  "paths": {
...
    "/pacientes": {
      "get": {
        "summary": "Listar pacientes",
        "description": "Lista todos los pacientes. Requiere rol de médico o admin.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      },
      "post": {
        "summary": "Crear paciente",
        "description": "Crea un nuevo paciente. Requiere rol de médico o admin.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PacienteCrear" }
            }
          }
        },
        "responses": {
          "201": { "description": "Paciente creado" },
          "400": { "description": "RUT o email ya existe" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
    "/pacientes/{id}": {
      "get": {
        "summary": "Obtener paciente",
        "description": "Obtiene la información de un paciente específico. Médicos y admins pueden ver cualquier paciente. Pacientes solo pueden ver sus propios datos.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Paciente no encontrado" }
        }
      },
      "put": {
        "summary": "Actualizar paciente",
        "description": "Actualiza la información de un paciente. Médicos y admins pueden actualizar cualquier paciente. Pacientes solo pueden actualizar sus propios datos.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PacienteActualizar" }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Error de validación" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Paciente no encontrado" }
        }
      },
      "delete": {
        "summary": "Eliminar paciente",
        "description": "Elimina un paciente. Solo admin puede eliminar pacientes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Paciente eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" },
          "404": { "description": "Paciente no encontrado" }
        }
      }
    },
...
    "/examenes/paciente/{paciente_id}": {
      "get": {
        "summary": "Obtener exámenes por paciente",
        "description": "Obtiene todos los exámenes médicos de un paciente específico. Médicos y admins pueden ver exámenes de cualquier paciente. Pacientes solo pueden ver sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "paciente_id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
...
    "/documentos/examen/{examen_id}": {
      "get": {
        "summary": "Listar documentos de un examen",
        "description": "Lista todos los documentos asociados a un examen médico. Médicos y admins pueden ver documentos de cualquier examen. Pacientes solo pueden ver documentos de sus propios exámenes.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "examen_id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "No autorizado" },
          "403": { "description": "No tiene permisos" }
        }
      }
    },
...
    "/tipo-examen/{id}": {
      "get": {
        "summary": "Obtener tipo de examen",
        "description": "Obtiene la información de un tipo de examen específico.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Tipo de examen no encontrado" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PacienteCrear": {
        "type": "object",
        "properties": {
          "rut": { "type": "string" },
          "primer_nombre": { "type": "string" },
          "apellido_paterno": { "type": "string" },
          "apellido_materno": { "type": "string" },
          "fecha_nacimiento": { "type": "string", "format": "date" },
          "sexo": { "type": "string", "enum": ["M", "F", "O"] },
          "nacionalidad": { "type": "string" },
          "estado_civil": { "type": "string" },
          "email": { "type": "string" },
          "telefono": { "type": "string" },
          "direccion": { "type": "string" },
          "contacto_emergencia_nombre": { "type": "string" },
          "contacto_emergencia_telefono": { "type": "string" },
          "contacto_emergencia_relacion": { "type": "string" },
          "tipo_prevision": { "type": "string" },
          "tiene_convenio": { "type": "boolean" },
          "grupo_sanguineo": { "type": "string" },
          "alergias": { "type": "string" },
          "enfermedades_cronicas": { "type": "string" },
          "discapacidades": { "type": "string" },
          "observaciones_medicas_generales": { "type": "string" },
          "usuario_id": { "type": "integer" }
        },
        "required": [
          "rut",
          "primer_nombre",
          "apellido_paterno",
          "fecha_nacimiento",
          "sexo",
          "email"
        ]
      },
      "PacienteActualizar": {
        "type": "object",
        "properties": {
          "rut": { "type": "string" },
          "primer_nombre": { "type": "string" },
          "apellido_paterno": { "type": "string" },
          "apellido_materno": { "type": "string" },
          "fecha_nacimiento": { "type": "string", "format": "date" },
          "sexo": { "type": "string", "enum": ["M", "F", "O"] },
          "nacionalidad": { "type": "string" },
          "estado_civil": { "type": "string" },
          "email": { "type": "string" },
          "telefono": { "type": "string" },
          "direccion": { "type": "string" },
          "contacto_emergencia_nombre": { "type": "string" },
          "contacto_emergencia_telefono": { "type": "string" },
          "contacto_emergencia_relacion": { "type": "string" },
          "tipo_prevision": { "type": "string" },
          "tiene_convenio": { "type": "boolean" },
          "grupo_sanguineo": { "type": "string" },
          "alergias": { "type": "string" },
          "enfermedades_cronicas": { "type": "string" },
          "discapacidades": { "type": "string" },
          "observaciones_medicas_generales": { "type": "string" },
          "usuario_id": { "type": "integer" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /autenticacion/login. Incluye: usuarioId, nombreUsuario, rolId, rolNombre, pacienteId, roles"
      }
    }
  }
}